rename: auth 도메인으로 디렉토리 구조 변경
